Add gool command to start minimal IDE
The gool "IDE" will hopefully make it easier for first-time programmers
to get up to speed. Learning how to use the command line, the go tool
and how to create a proper directory structure for code and tools was
very time consuming. With this we can hopefully focus on the code.